Tablet supports rotation.
When the tablet is rotated, the selected workout doesn't go back to
the first workout in the list anymore because we save the state of
the workout so that when the fragment along with the activity is
destroyed and recreated during a rotation, the selected fragment is
shown again.

Remove unnecessary casting as reported by the code analyzer.