Ch. 11: Nest stopwatch fragment inside of detail fragment.
Now, we're back to the workout list as the main page, and the
stopwatch is a nested fragment in the detail fragment. Since
getSupportedFragmentManager() return's the fragment's parent activity,
we use getChildFragmentManager() to get the fragment's parent fragment.
This nests the transactions of stopwatch fragment inside detail
fragment, whereas getFragmentManager() would put the stopwatch fragment
transaction on top of the detail fragment on the back stack. This
nesting vs. layering on top allows a single back button press to get
back to the workout list as opposed to two back button clicks to get
back to the workout list respectively.